locationsResolver
